drop exists

+++ database/migrations/2016_05_26_064007_create_employees_table.php
@@ -65,6 +65,8 @@ class CreateEmployeesTable extends Migration
      */
     public function down()
     {
-        Schema::drop('employees');
+        if (Schema::hasTable('employees')) {
+            Schema::drop('employees');
+        }
     }
 }
